:recycle: seperate timer, swipe logic to custom hook
